+++ crates/core/tests/java/do_not_indent_top_level_class_members/members.out.java
@@ -1,0 +1,13 @@
+class TopLevel {
+int field;
+
+void method() {
+    int local = 1;
+}
+
+class Nested {
+    int nestedField;
+
+    void nestedMethod() {}
+}
+}