+++ crates/core/tests/java/keep_indents_on_empty_lines/blank_lines_default.out.java
@@ -1,0 +1,17 @@
+class BlankLines {
+    void method() {
+        int a = 1;
+
+        int b = 2;
+
+        if (a > b) {
+            int c = 3;
+
+            int d = 4;
+        }
+    }
+
+    void other() {
+        int x = 1;
+    }
+}